Style: more cleanup, move turret to more adequate place
