+++ src/main/java/frc/robot/subsystems/turret/flywheels/FlywheelConstants.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.flywheel;
+package frc.robot.subsystems.turret.flywheels;
 
 import edu.wpi.first.math.system.plant.DCMotor;
 import frc.lib.generic.hardware.motor.*;
